README update & Code cleanup

+++ src/net/minecraft/server/DataWatcher.java
@@ -21,30 +21,30 @@ public class DataWatcher {
             throw new IllegalArgumentException("Unknown data type: " + object.getClass());
         } else if (i > 31) {
             throw new IllegalArgumentException("Data value id is too big with " + i + "! (Max is " + 31 + ")");
-        } else if (this.b.containsKey(Integer.valueOf(i))) {
+        } else if (this.b.containsKey(i)) {
             throw new IllegalArgumentException("Duplicate id value for " + i + "!");
         } else {
             WatchableObject watchableobject = new WatchableObject(integer.intValue(), i, object);
 
-            this.b.put(Integer.valueOf(i), watchableobject);
+            this.b.put(i, watchableobject);
             this.d = false;
         }
     }
 
     public byte a(int i) {
-        return ((Byte) ((WatchableObject) this.b.get(Integer.valueOf(i))).b()).byteValue();
+        return ((Byte) ((WatchableObject) this.b.get(i)).b()).byteValue();
     }
 
     public int b(int i) {
-        return ((Integer) ((WatchableObject) this.b.get(Integer.valueOf(i))).b()).intValue();
+        return ((Integer) ((WatchableObject) this.b.get(i)).b()).intValue();
     }
 
     public String c(int i) {
-        return (String) ((WatchableObject) this.b.get(Integer.valueOf(i))).b();
+        return (String) ((WatchableObject) this.b.get(i)).b();
     }
 
     public void watch(int i, Object object) {
-        WatchableObject watchableobject = (WatchableObject) this.b.get(Integer.valueOf(i));
+        WatchableObject watchableobject = (WatchableObject) this.b.get(i);
 
         if (!object.equals(watchableobject.b())) {
             watchableobject.a(object);
@@ -163,19 +163,19 @@ public class DataWatcher {
 
             switch (i) {
             case 0:
-                watchableobject = new WatchableObject(i, j, Byte.valueOf(datainputstream.readByte()));
+                watchableobject = new WatchableObject(i, j, datainputstream.readByte());
                 break;
 
             case 1:
-                watchableobject = new WatchableObject(i, j, Short.valueOf(datainputstream.readShort()));
+                watchableobject = new WatchableObject(i, j, datainputstream.readShort());
                 break;
 
             case 2:
-                watchableobject = new WatchableObject(i, j, Integer.valueOf(datainputstream.readInt()));
+                watchableobject = new WatchableObject(i, j, datainputstream.readInt());
                 break;
 
             case 3:
-                watchableobject = new WatchableObject(i, j, Float.valueOf(datainputstream.readFloat()));
+                watchableobject = new WatchableObject(i, j, datainputstream.readFloat());
                 break;
 
             case 4:
@@ -209,12 +209,12 @@ public class DataWatcher {
     }
 
     static {
-        a.put(Byte.class, Integer.valueOf(0));
-        a.put(Short.class, Integer.valueOf(1));
-        a.put(Integer.class, Integer.valueOf(2));
-        a.put(Float.class, Integer.valueOf(3));
-        a.put(String.class, Integer.valueOf(4));
-        a.put(ItemStack.class, Integer.valueOf(5));
+        a.put(Byte.class, 0);
+        a.put(Short.class, 1);
+        a.put(Integer.class, 2);
+        a.put(Float.class, 3);
+        a.put(String.class, 4);
+        a.put(ItemStack.class, 5);
         a.put(ChunkCoordinates.class, Integer.valueOf(6));
     }
 }
